Team_recommendation_Ai/requirement.txt
fastapi
uvicorn
psycopg2-binary
pydantic
google-generativeai